requirements.txt
Django==1.11.18
djangorestframework==3.9.0
djangorestframework-simplejwt==3.3
django-rest-swagger==2.1.2
psycopg2==2.7.7
psycopg2-binary==2.7.7